+++ deliveryService/src/main/java/edu/tum/ase/deliveryService/service/DeliveryService.java
@@ -1,10 +1,15 @@
 package edu.tum.ase.deliveryService.service;
 
+import edu.tum.ase.deliveryService.DeliveryServiceApplication;
 import edu.tum.ase.deliveryService.exceptions.BoxAlreadyExistsException;
 import edu.tum.ase.deliveryService.exceptions.DeliveryNotFoundException;
+import edu.tum.ase.deliveryService.exceptions.DeliveryStatusException;
 import edu.tum.ase.deliveryService.model.Box;
 import edu.tum.ase.deliveryService.model.Delivery;
+import edu.tum.ase.deliveryService.repository.BoxRepository;
 import edu.tum.ase.deliveryService.repository.DeliveryRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,8 +19,13 @@ import java.util.List;
 @Service
 public class DeliveryService {
 
+    private static final Logger log = LoggerFactory.getLogger(DeliveryServiceApplication.class);
+
     @Autowired
     private DeliveryRepository deliveryRepository;
+
+    @Autowired
+    private BoxRepository boxRepository;
 
     //##################################################################################################################
     // Create
@@ -43,6 +53,7 @@ public class DeliveryService {
     // Update
 
     public Delivery updateDelivery(Delivery delivery) {
+        log.info("Updating delivery " + delivery);
         return deliveryRepository.save(delivery);
     }
 
@@ -50,6 +61,15 @@ public class DeliveryService {
     // Delete
 
     public void deleteDelivery(Delivery delivery) {
+        if (!delivery.getStatus().canBeRemoved()) {
+            throw new DeliveryStatusException();
+        }
+
+        Box box = delivery.getBox();
+        box.removeDelivery(delivery);
+        boxRepository.save(box);
+
+        log.info("Deleting delivery " + delivery);
         deliveryRepository.delete(delivery);
     }
 }
